% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DU_SampleAges}
\alias{DU_SampleAges}
\title{Sample Age Data}
\format{
A data frame with 
\describe{
  \item{DU}{ DU Name}
  \item{Avg_Gen}{Average generation time (i.e. dominant age class}
  \item{}{}
  ...
}
}
\source{
Dummy Data
}
\usage{
DU_SampleAges
}
\description{
A dataset containing dominant age classes for the DUs in DU_SampleData
}
\keyword{datasets}
