% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DU_SampleData}
\alias{DU_SampleData}
\title{Sample Abundance Data}
\format{
A data frame with  5 required and some optional variables:
\describe{
  \item{Species}{ Species}
  \item{Year}{ estimate year}
  \item{DU}{ DU Name}
  \item{Abd}{abundance estimate}
  \item{}{}
  ...
}
}
\source{
Dummy Data
}
\usage{
DU_SampleData
}
\description{
A dataset containing abundance data (may be total abundance, or a relative index) for various Sockeye, Coho, and Chinook stocks.
}
\keyword{datasets}
