% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_multiFit.R
\name{multiFit}
\alias{multiFit}
\title{calcPercChangeSimple}
\usage{
multiFit(data.df, window.df, plot.file = "PercChange_Plots.pdf")
}
\arguments{
\item{data.df}{data frame with columns DU, Year, Abd}

\item{window.df}{data frame with time windows for each DU (e.g. 3 gen +1)}

\item{plot.file}{text string with path and filename for pdf of diagnostic plots}
}
\description{
this function applies the function \code{\link{comparePercChange}} to each DU in a data set and generates summary outputs plus diagnostic plots.
}
