% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStan.R
\name{run.stan}
\alias{run.stan}
\title{Function to run stan to esimate probabiltiy of declines
Created by: Carrie Holt
params list for function params}
\usage{
run.stan(
  du.label,
  du.df,
  yrs.window,
  perc.change.bm = c(-30, -50, -70),
  calc.yr,
  standardize.data = TRUE,
  scenario.name = "test",
  path.results = "",
  mcmc.plots = TRUE,
  prior_sigma_type = "exp",
  prior_sigma = 2.5,
  H0 = FALSE
)
}
\arguments{
\item{du.label}{[character()] # pathing for the project}

\item{du.df}{[data.frame()] input data}

\item{yrs.window}{[numeric()] [vector()] # ? - guessing at from Larkin}

\item{perc.change.bm}{[character()] [vector()] # ? - guessing at from Larkin}

\item{calc.yr}{[numeric()]}

\item{standardize.data}{[logical()]}

\item{scenario.name}{[character()]}

\item{path.results}{[character()]}

\item{mcmc.plots}{[logical()]}

\item{prior_sigma_type}{[character()]}

\item{prior_sigma}{[numeric()]}

\item{H0}{[logical()]}
}
\description{
Function to run stan to esimate probabiltiy of declines
Created by: Carrie Holt
params list for function params
}
