% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_func.R
\name{testing_fun}
\alias{testing_fun}
\title{Test function for working on Stan inclusion
params list for function params
Testing out load_all() devtools and roxygenise}
\usage{
testing_fun(arg1 = "Change me")
}
\arguments{
\item{arg1}{A value to return}
}
\description{
Test function for working on Stan inclusion
params list for function params
Testing out load_all() devtools and roxygenise
}
\examples{
testing_fun("sample") # returns "sample"
}
